Atualização e alteração das constantes, de int para byte.

+++ src/main/java/br/com/luisfga/talkingz/server/database/sql/dao/MessageDAO.java
@@ -11,7 +11,7 @@ import br.com.luisfga.talkingz.server.database.sql.entity.Message;
 import io.agroal.api.AgroalDataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import br.com.luisfga.talkingz.commons.MessageStatus;
+import br.com.luisfga.talkingz.commons.constants.MessageStatus;
 import br.com.luisfga.talkingz.commons.MessageWrapper;
 import br.com.luisfga.talkingz.commons.orchestration.command.CommandSend;
 
@@ -100,8 +100,8 @@ public class MessageDAO {
     }
 
     @Transactional
-    public void updateMessageStatus(UUID uuid, int statusMsg){
+    public void updateMessageStatus(UUID uuid, byte statusMsg){
         Message message = em.find(Message.class, uuid);
-        message.msgStatus = (short) statusMsg;
+        message.msgStatus = statusMsg;
     }
 }
